+++ docs/md/releasenotes.md
@@ -15,6 +15,7 @@
 - Add new `InterfaceMethodReferenceFixer` that patches `MethodRefConstant`s to `InterfaceMethodRefConstant`s in visited classes when the reference is an interface method.
 - Add `reset` method to `MemberRemover`.
 - Add processing flags helper methods to `Processable`.
+- Add `isLeader` and `isFallThrough` methods to the `BranchTargetFinder`. These methods can be used to determine if an offset is the leader of a basic block or the fallthrough of a conditional branch respectively.
 
 ### Improved